visual point test works with ReLu

+++ src/backend/src/FileManager.cpp
@@ -61,7 +61,7 @@ void FileManager::to_file(neural_network::NeuralNetwork& net){
 
             auto Neuron = hiddenLayer->neurons[neuron];
             writer << Neuron->bias << ' ';
-            for (int weight = 0; weight < Neuron->weights.size(); weight++){
+            for (size_t weight = 0; weight < Neuron->weights.size(); weight++){
                 writer << Neuron->weights[weight] << ' ';
             }
             writer << '\n';
@@ -72,7 +72,7 @@ void FileManager::to_file(neural_network::NeuralNetwork& net){
     for(int neuron = 0; neuron < netStruct->out->node_out; neuron++){
         auto Neuron = netStruct->out->neurons[neuron];
         writer << Neuron->bias << ' ';
-        for (int weight = 0; weight < Neuron->weights.size(); weight++){
+        for (size_t weight = 0; weight < Neuron->weights.size(); weight++){
             writer << Neuron->weights[weight] << ' ';
         }
         writer << '\n';
